Splitted CAN hard drivers from processing logic
