+++ include/terminalcommands.h
@@ -18,7 +18,7 @@
  */
 #ifndef TERMINALCOMMANDS_H
 #define TERMINALCOMMANDS_H
-
+#include "canmap.h"
 
 class TerminalCommands
 {
@@ -32,11 +32,13 @@ class TerminalCommands
       static void SaveParameters(Terminal* term, char *arg);
       static void LoadParameters(Terminal* term, char *arg);
       static void Reset(Terminal* term, char *arg);
+      static void SetCanMap(CanMap* m) { canMap = m; }
 
    protected:
 
    private:
       static void PrintCanMap(Param::PARAM_NUM param, uint32_t canid, uint8_t offset, uint8_t length, float gain, bool rx);
+      static CanMap* canMap;
 };
 
 #endif // TERMINALCOMMANDS_H
